Small internal updates
Instead of adding LunarROMCorruptor every time to error messages, it now references the program name, so changing the program name is easier now.
Catched invalid process operation
Added check to ensure that files in the corruption queue are sent to the corruptor first before corrupting, before it used to just barrel through.
Changed wording on a few messages.
Added "Randomize Corruption Engine" to the Automation tab.